Move to TypeScript / TSDX (#51)
Changes:

- Move to TypeScript using TSDX.
- Dynamically import connectors.
- Use provider.request() (EIP 1193) when possible.
- Remove the SquareLink connector.
- Move to Prettier 2.
