--- .babelrc.json
@@ -1,12 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "targets": "> 2%, not dead, not ie > 0",
-        "modules": false
-      }
-    ],
-    "@babel/preset-react"
-  ]
-}
