+++ .gitignore
@@ -1,4 +1,5 @@
-node_modules/
-.cache/
 dist/
+node_modules/
 yarn-error.log
+.yarn/cache
+.cache/
