+++ examples/simple-connect/package.json
@@ -4,15 +4,19 @@
     "build": "parcel build index.html"
   },
   "dependencies": {
-    "babel-polyfill": "^6.26.0",
-    "ethers": "^4.0.43",
-    "jsbi": "^3.1.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
+    "jsbi": "^3.1.4",
+    "react": "link:../../node_modules/react",
+    "react-dom": "link:../../node_modules/react-dom",
+    "token-amount": "^0.1.0",
     "use-wallet": "link:../.."
   },
+  "alias": {
+    "react": "../../node_modules/react",
+    "react-dom": "../../node_modules/react-dom"
+  },
   "devDependencies": {
-    "@babel/core": "^7.8.3",
+    "@babel/core": "^7.12.9",
+    "@babel/preset-react": "^7.10.4",
     "parcel": "^1.12.4"
   }
 }
