+++ package.json
@@ -4,60 +4,66 @@
   "author": "Aragon Association <legal@aragon.org>",
   "license": "MIT",
   "main": "dist/cjs/index.js",
-  "module": "dist/index.js",
+  "module": "dist/esm/index.js",
+  "types": "dist/esm/index.d.ts",
   "sideEffects": false,
-  "types": "types/index.d.ts",
   "files": [
-    "dist",
-    "types/index.d.ts"
+    "dist"
   ],
   "scripts": {
-    "build": "yarn rollup --config",
-    "dev": "yarn rollup --config --watch",
-    "lint": "yarn eslint ./src",
-    "test": "yarn lint",
-    "prepare": "rm -rf ./dist && yarn build",
+    "start": "tsdx watch",
+    "build": "tsdx build",
+    "test": "tsdx test --passWithNoTests",
+    "lint": "tsdx lint src",
     "prepublishOnly": "git push && git push --tags"
   },
   "dependencies": {
-    "@aragon/provided-connector": "^6.0.7",
-    "@web3-react/authereum-connector": "^6.1.1",
+    "@aragon/provided-connector": "^6.0.8",
+    "@typescript-eslint/parser": "^4.1.0",
     "@web3-react/core": "^6.1.1",
-    "@web3-react/fortmatic-connector": "^6.0.9",
+    "@web3-react/fortmatic-connector": "^6.1.6",
     "@web3-react/frame-connector": "^6.0.9",
     "@web3-react/injected-connector": "^6.0.7",
-    "@web3-react/portis-connector": "^6.1.1",
-    "@web3-react/squarelink-connector": "^6.1.1",
-    "@web3-react/torus-connector": "^6.1.1",
-    "@web3-react/walletconnect-connector": "^6.1.4",
-    "@web3-react/walletlink-connector": "^6.1.1",
-    "jsbi": "^3.1.1"
+    "@web3-react/portis-connector": "^6.1.6",
+    "@web3-react/torus-connector": "^6.1.7",
+    "@web3-react/walletconnect-connector": "^6.1.6",
+    "@web3-react/walletlink-connector": "^6.1.6",
+    "jsbi": "^3.1.4",
+    "prop-types": "^15.7.2",
+    "regenerator-runtime": "^0.13.7"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "tsdx lint"
+    }
   },
   "devDependencies": {
-    "@babel/cli": "^7.8.3",
-    "@babel/core": "^7.8.3",
-    "@babel/preset-env": "^7.8.3",
-    "@babel/preset-react": "^7.8.3",
-    "@rollup/plugin-commonjs": "^11.0.1",
-    "@rollup/plugin-json": "^4.0.1",
-    "@rollup/plugin-node-resolve": "^7.0.0",
-    "babel-eslint": "^10.0.3",
-    "builtin-modules": "^3.1.0",
-    "eslint": "^6.8.0",
+    "@types/react": "^16.9.49",
+    "@types/react-dom": "^16.9.8",
+    "@typescript-eslint/eslint-plugin": "^4.1.0",
+    "@web3-react/abstract-connector": "^6.0.7",
+    "babel-eslint": "^10.1.0",
+    "eslint": "^7.8.1",
     "eslint-config-prettier": "^6.9.0",
+    "eslint-config-react-app": "^5.2.1",
     "eslint-config-standard": "^14.1.0",
     "eslint-config-standard-react": "^9.2.0",
+    "eslint-plugin-flowtype": "^5.2.0",
     "eslint-plugin-import": "^2.20.0",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
     "eslint-plugin-node": "^11.0.0",
     "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.18.0",
-    "eslint-plugin-react-hooks": "^2.3.0",
+    "eslint-plugin-react": "^7.20.6",
+    "eslint-plugin-react-hooks": "^4.1.0",
     "eslint-plugin-standard": "^4.0.1",
-    "prettier": "^1.19.1",
-    "rollup": "^1.29.1",
-    "rollup-plugin-babel": "^4.3.3",
-    "rollup-plugin-terser": "^5.2.0"
+    "husky": "^4.3.0",
+    "prettier": "^2.1.1",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "tsdx": "^0.14.1",
+    "tslib": "^2.0.1",
+    "typescript": "^4.0.2"
   },
   "peerDependencies": {
     "react": "^16.12.0"
